+++ app/containers/MainContainer.tsx
@@ -5,7 +5,7 @@ import Contact from '../components/Contact';
 import Settings from '../components/Settings';
 import Occupied from '../components/Occupied/Occupied';
 import lightAndDark from '../components/Styling';
-import GraphsContainer from './GraphsContainer';
+import GraphsContainer from './GraphsContainer/GraphsContainer';
 import { DashboardContext } from '../context/DashboardContext';
 import SignUp from '../components/SignUp';
 import Login from '../components/Login';
